requirements.txt
alabaster==0.7.12
astroid==2.4.2
asyncio==3.4.3
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
docopt==0.6.2
docutils==0.16
Flask==1.1.2
Flask-Ext==0.1
Flask-PyMongo==2.3.0
future==0.18.2
hbmqtt==0.9.6
idna==2.10
imagesize==1.2.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongotriggers==2.0.1
motor==2.3.0
netifaces==0.10.9
packaging==20.4
paho-mqtt==1.5.1
passlib==1.7.4
Pygments==2.7.2
pylint==2.6.0
pymongo==3.11.0
PyMySQL==0.10.1
pyngrok==4.2.2
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
toml==0.10.1
transitions==0.8.4
typed-ast==1.4.1
urllib3==1.25.10
websockets==8.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
